requirements.txt
streamlit
requests
python-dotenv
pydub
mutagen
tenacity
elevenlabs
audioop-lts
pydantic
